debian/gitweb: avoid trigger loop by using noawait triggers
The intent was to prevent the gitweb from satisfying dependencies
until the trigger runs.  Triggers don't work that way --- it is the
triggering package (apache2 in this case) that awaits the trigger run
before satisfying dependencies, which is counterproductive in this
example.

Use the -noawait variant instead.  In practice, no one is using a
dependency on the gitweb package to imply an ordering in webapp startup,
so we can do without the protection that we were hoping would come from
delaying the 'installed' state.

Without this change, upgrades often fail with

 dpkg: cycle found while processing triggers:
  chain of packages whose triggers are or may be responsible:
   gitweb -> gitweb
  packages' pending triggers which are or may be unresolvable:
   gitweb: /usr/share/apache2/apache2-maintscript-helper

Signed-off-by: Jonathan Nieder <jrnieder@gmail.com>